cleanup: adjust folder structure in preparation for modularization
