+++ app/(dashboard)/(routes)/quiz/database/_components/columns.tsx
