+++ app/(dashboard)/(routes)/quiz/database/_components/data-table.tsx
